Updates to processing of image url's for user pictures
Since July's versioning number request, JSON replies contain a prefix
and a suffix for pictures. These need to be concatenated with a size
(like 32/48/64 or original) in between.

Before, when no version paramater was given, replies would be with the
oldest API, now they get the reply back from the newest API.